# src/main/resources/source/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: 2294678765@qq.com
    name: API Support
    url: https://github.com/FranklinThree
  description: 基于SpringBoot的响应式技术博客的设计和实现前后端接口实现
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: frontend-backend-api
  version: "1.0"
servers:
- url: http://localhost:9090
...
paths:
...
  /user/register:
    post:
      description: 用户注册
      operationId: post-user-register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/post_user_register_request'
        description: ""
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: 用户注册成功
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: 参数错误
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: 邮箱或用户名已被使用
      servers:
      - url: http://localhost:9090
      summary: User Register
      tags:
      - public
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: public
  /user/login:
    post:
      description: 用户登录
      operationId: post-user-login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/post_user_login_request'
        description: email和username填一项即可
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post_user_login_202_response'
          description: 用户登录成功
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: 参数错误
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: 用户登录失败，密码或用户名错误
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
          description: 账号不可用
      servers:
      - url: http://localhost:9090
      summary: User Login
      tags:
      - public
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: public
...
components:
  schemas:
    User:
      description: 一般用户
      example:
        realName: realName
        emailVerified: true
        password: password
        dateOfBirth: 1997-10-31T00:00:00.000+00:00
        id: id
        email: email
        username: username
        createDate: 2000-01-23
        status: null
      properties:
        id:
          description: 唯一标识符
          type: string
        username:
          description: 用户名
          type: string
        realName:
          description: 真实姓名
          type: string
        email:
          description: 邮箱
          format: email
          type: string
        dateOfBirth:
          description: 生日
          example: 1997-10-31
          format: date
          type: string
        emailVerified:
          description: Set to true if the user's email has been verified.
          type: boolean
        createDate:
          description: The date that the user was created.
          format: date
          type: string
        status:
          $ref: '#/components/schemas/AccountStatus'
        password:
          description: 密码
          type: string
      title: User
      type: object
    AccountStatus:
      description: 账户状态
      enum:
      - regular
      - frozen
      - disabled
      title: AccountStatus
      type: string
    ErrorInfo:
      description: 错误信息
      properties:
        id:
          description: 唯一标识符
          type: string
        code:
          description: 错误码
          type: integer
        description:
          description: 说明
          type: string
        given:
          description: 现有的情况
          type: string
        expected:
          description: 期望的情况
          type: string
      title: ErrorInfo
      type: object
...
    post_user_register_request:
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
      - user
      type: object
    post_user_login_request:
      properties:
        email:
          description: 邮箱
          type: string
        username:
          description: 用户名
          type: string
        password:
          description: 密码
          type: string
      required:
      - password
      type: object
    post_user_login_202_response:
      example:
        accessToken: accessToken
        user:
          realName: realName
          emailVerified: true
          password: password
          dateOfBirth: 1997-10-31T00:00:00.000+00:00
          id: id
          email: email
          username: username
          createDate: 2000-01-23
          status: null
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          description: 访问令牌
          type: string
      required:
      - accessToken
      - user
      type: object
...